pyproject.toml
[project]
name = "simple-fastapi-project"
version = "0.1.0"
description = "Simple FastAPI project template"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "asyncpg>=0.31.0",
    "fastapi>=0.135.1",
    "pydantic[email]>=2.12.5",
    "pyjwt>=2.11.0",
    "python-multipart>=0.0.22",
    "sqlalchemy[asyncio]>=2.0.48",
    "uvicorn>=0.41.0",
]

[dependency-groups]
dev = [
    "pre-commit>=4.5.1",
    "pytest>=9.0.2",
    "pytest-cov>=7.0.0",
    "pytest-env>=1.5.0",
    "python-dotenv>=1.2.2",
    "ruff>=0.15.5",
    "testcontainers>=4.14.1",
]
